@@ -1,9 +1,6 @@
 package com.example.zooseeker_t62;
 
-import android.app.AlertDialog;
-import android.app.Instrumentation;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -14,15 +11,14 @@ import android.widget.CompoundButton;
 import android.widget.TextView;
 import android.widget.ToggleButton;
 
+import androidx.annotation.VisibleForTesting;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModelProvider;
 //import androidx.test.core.app.ApplicationProvider;
 //import androidx.test.platform.app.InstrumentationRegistry;
 
 
 import java.util.ArrayList;
-import java.util.Locale;
 import java.util.Map;
 
 import org.jgrapht.Graph;
@@ -30,10 +26,9 @@ import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 
 import java.util.List;
-import java.util.Observable;
-import java.util.Observer;
-import java.util.Set;
 import java.util.Stack;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
 
 /**
  * @description: Uses our algorithm to calculate optimal route of exhibit paths
@@ -64,6 +59,7 @@ public class RouteDirectionsActivity extends AppCompatActivity {
     private List<ExhibitItem> exhibits;
     private List<ExhibitItem> unvisited;
     private Stack<ExhibitItem> visited;
+    private LocationModel model;
 
 
     @Override
@@ -71,13 +67,17 @@ public class RouteDirectionsActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.route_directions);
 
+        model = new ViewModelProvider(this).get(LocationModel.class);
+        //var locationManager = (LocationManager) this.getSystemService(Context.LOCATION_SERVICE);
+        //var provider = LocationManager.GPS_PROVIDER;
+        //model.addLocationProviderSource(locationManager, provider);
         exhibits = getPlannerExhibits();
 
         //Log.d("RouteDirectionsActivity.java onCreate", exhibits.toString());
 
         loadGraphData();
 
-        List<ExhibitItem> allExhibits = ExhibitItem.loadJSON(this, "sample_ms1_demo_node_info.json");
+        List<ExhibitItem> allExhibits = ExhibitItem.loadJSON(this, "sample_ms2_exhibit_info.json");
 
         unvisited = new ArrayList<>();
 
@@ -188,6 +188,9 @@ public class RouteDirectionsActivity extends AppCompatActivity {
 
         String prevNode = visited.peek().id;
         if (prevNode.equals(currNode)) {
+            if (visited.size() == 1) {
+                return false;
+            }
             visited.pop();
             prevNode = visited.peek().id;
         }
@@ -238,9 +241,11 @@ public class RouteDirectionsActivity extends AppCompatActivity {
      * @description: find exhibit distance from entrance
      */
     public static String findExhibitDist(Context context, String entrance, String id) {
-        Graph<String, IdentifiedWeightedEdge> g = ZooData.loadZooGraphJSON("sample_ms1_demo_zoo_graph.json", context);
-        Map<String, ZooData.VertexInfo> vInfo = ZooData.loadVertexInfoJSON("sample_ms1_demo_node_info.json", context);
-        Map<String, ZooData.EdgeInfo> eInfo = ZooData.loadEdgeInfoJSON("sample_ms1_demo_edge_info.json", context);
+        Graph<String, IdentifiedWeightedEdge> g = ZooData.loadZooGraphJSON("sample_ms2_zoo_graph.json", context);
+        Map<String, ZooData.VertexInfo> vInfo = ZooData.loadVertexInfoJSON("sample_ms2_exhibit_info.json", context);
+        Map<String, ZooData.EdgeInfo> eInfo = ZooData.loadEdgeInfoJSON("sample_ms2_trail_info.json", context);
+
+        Log.d("RouteDirectionsActivity", "entrance: " + entrance + " id: " + id);
 
         GraphPath<String, IdentifiedWeightedEdge> path = DijkstraShortestPath.findPathBetween(g, entrance, id);
 
@@ -256,9 +261,9 @@ public class RouteDirectionsActivity extends AppCompatActivity {
      */
     public boolean loadGraphData() {
         try {
-            g = ZooData.loadZooGraphJSON("sample_ms1_demo_zoo_graph.json", this);
-            vInfo = ZooData.loadVertexInfoJSON("sample_ms1_demo_node_info.json", this);
-            eInfo = ZooData.loadEdgeInfoJSON("sample_ms1_demo_edge_info.json", this);
+            g = ZooData.loadZooGraphJSON("sample_ms2_zoo_graph.json", this);
+            vInfo = ZooData.loadVertexInfoJSON("sample_ms2_exhibit_info.json", this);
+            eInfo = ZooData.loadEdgeInfoJSON("sample_ms2_trail_info.json", this);
         }
         catch (Exception e) {
             return false;
@@ -332,7 +337,7 @@ public class RouteDirectionsActivity extends AppCompatActivity {
      * we simply decrement pathIdx and thus the previous path string will display
      */
     public void onPrevClick(View view) {
-        List<ExhibitItem> allExhibits = ExhibitItem.loadJSON(this, "sample_ms1_demo_node_info.json");
+        List<ExhibitItem> allExhibits = ExhibitItem.loadJSON(this, "sample_ms2_exhibit_info.json");
         if (!calcPrevStep()) {
 //            Log.d("test", "returns false");
             Intent intent = new Intent(this, ExhibitActivity.class);
@@ -411,11 +416,38 @@ public class RouteDirectionsActivity extends AppCompatActivity {
         startActivity(intent);
     }
 
+    public void onCoordUpdateClick(View view) {
+        //Log.d("onCoordUpdateClick()", view.toString());
+        TextView coordsText = findViewById(R.id.coords_edit_txt);
+
+        Log.d("onCoordUpdateClick()", coordsText.getText().toString());
+        String[] coords = coordsText.getText().toString().split(",");
+        double coordLat = Double.parseDouble(coords[0]);
+        double coordLong = Double.parseDouble(coords[1]);
+
+        Log.d("onCoordUpdateClick()", "latitude: " + coordLat + ", longitude: " + coordLong);
+
+        Coord updatedCoords = new Coord(coordLat, coordLong);
+        mockLocation(updatedCoords);
+
+        model.getLastKnownCoords().observe(this, (coord) -> Log.i("onCoordUpdateClick", String.format("Observing location model update to %s", coord)));
+    }
+
     /**
      * @description: Proper lifecycle cleanup
      */
     @Override
     protected void onDestroy() {
         super.onDestroy();
+    }
+
+    @VisibleForTesting
+    public void mockLocation(Coord coords) {
+        model.mockLocation(coords);
+    }
+
+    @VisibleForTesting
+    public Future<?> mockRoute(List<Coord> route, long delay, TimeUnit unit) {
+        return model.mockRoute(route, delay, unit);
     }
 }
